+++ src/providers/gemini-provider.ts
@@ -10,7 +10,7 @@
 import type { Env } from '../types/env'
 import type { BookshelfDetectedBook } from '../types/gemini-schemas'
 import { BOOKSHELF_RESPONSE_SCHEMA } from '../types/gemini-schemas'
-import { retryWithBackoff } from '../utils/retry'
+import { retryWithBackoff } from '../utils/concurrency/retry'
 
 // Configurable model selection (Issue #101: Flash-Lite migration)
 // Override via GEMINI_VISION_MODEL env var if needed (rollback: set to "gemini-2.5-flash")
